NCBC/pyproject.toml
[project]
name = "ncbc"
version = "0.1.0"
description = ""
authors = [{ name = "FUJISHIGE TEMMA", email = "tenma.x0@gmail.com" }]
dependencies = [
    "click~=8.1.3",
    "matplotlib~=3.7.1",
    "numpy~=1.25.1",
    "scipy~=1.11.0",
    "h5py>=3.9.0",
    "neuron>=8.2.2",
    "typeid-python>=0.2.1",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff~=0.0.270",
    "ipykernel~=6.24.0",
    "pip~=23.1.2",
    "black>=23.7.0",
    "setuptools>=68.1.0",
]

[tool.ruff]
line-length = 180
exclude = [".git", ".venv", "__pycache__"]

[tool.hatch.metadata]
allow-direct-references = true
